Improve docs and add developer guide
