+++ app/scripts/spotlight/index.ts
@@ -1,0 +1,1 @@
+export { initSpotlight, closeSpotlight } from './ui';
