+++ app/scripts/spotlight/types.ts
@@ -1,0 +1,36 @@
+export interface Command {
+  id: string;
+  category: string;
+  title: string;
+  icon?: string;
+}
+
+export interface EntityInfo {
+  logicalName: string;
+  displayName: string;
+  primaryIdAttribute: string;
+  primaryNameAttribute: string;
+  logicalCollectionName: string;
+}
+
+export interface UserInfo {
+  userId: string;
+  userName: string;
+  fullName: string;
+}
+
+export enum Step {
+  Commands,
+  OpenRecordEntity,
+  OpenRecordId,
+  ImpersonateSearch,
+  FetchXml,
+  EntityInfoDisplay,
+}
+
+export interface SpotlightState {
+  query: string;
+  state: Step;
+  pills: string[];
+  selectedEntity: string;
+}
